+++ pkg/template/worker_template.go
@@ -221,7 +221,7 @@ systemd:
         --container-runtime=remote \
         --container-runtime-endpoint=unix:///run/containerd/containerd.sock \
         --logtostderr=true \
-        {{- if eq .Cluster.Kubernetes.CloudProvider "aws" }}
+        {{- if eq .ExternalCloudControllerManager }}
         --cloud-provider=external \
         {{ else -}}
         --cloud-provider={{.Cluster.Kubernetes.CloudProvider}} \